Isolate redis-entraid dependency for tests (#3521)
